Renames example plugin folders based on convention
Snap plugin naming convention is snap-plugin-<type>-<name>.
