+++ examples/snap-plugin-collector-rand/README.md
