+++ examples/snap-plugin-collector-rand/rand/rand_test.go
