+++ examples/snap-plugin-processor-reverse/reverse/reverse.go
