+++ examples/snap-plugin-processor-reverse/reverse/reverse_test.go
