+++ examples/snap-plugin-publisher-file/README.md
