+++ examples/snap-plugin-publisher-file/file/file_test.go
